[Feat] team user api 연결, 컴포넌트 분리
